Move: 폴더 구조 개편
파일이 조금 많아져서 보기 편하게 폴더별로 분류 하였습니다.

각 관심사 별로 소스와 test파일을 묶었습니다. 그리고 나머지는 'app'이라는 폴더를 만들어서 컴포넌트와 관계없는 것들을 모았습니다.
