+++ src/Restaurants/Restaurants.test.jsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import state from '../fixtures/state';
+import state from '../../fixtures/state';
 
 import Restaurants from './Restaurants';
